+++ src/AggregateFunctions/AggregateFunctionUniq.h
@@ -17,16 +17,15 @@
 #include <Interpreters/AggregationCommon.h>
 
 #include <Common/CombinedCardinalityEstimator.h>
-#include <Common/CurrentThread.h>
 #include <Common/HashTable/Hash.h>
 #include <Common/HashTable/HashSet.h>
 #include <Common/HyperLogLogWithSmallSetOptimization.h>
 #include <Common/assert_cast.h>
-#include <Common/setThreadName.h>
 #include <Common/typeid_cast.h>
 
 #include <AggregateFunctions/IAggregateFunction.h>
 #include <AggregateFunctions/ThetaSketchData.h>
+#include <AggregateFunctions/UniqExactSet.h>
 #include <AggregateFunctions/UniqVariadicHash.h>
 #include <AggregateFunctions/UniquesHashSet.h>
 
@@ -38,118 +37,6 @@ struct Settings;
 namespace ErrorCodes
 {
     extern const int NOT_IMPLEMENTED;
-}
-
-namespace detail
-{
-    enum class HTKind
-    {
-        SingleLevel,
-        TwoLevel,
-        Unknown
-    };
-
-    template <typename SingleLevelSet, typename TwoLevelSet>
-    class Set
-    {
-        static_assert(std::is_same_v<typename SingleLevelSet::value_type, typename TwoLevelSet::value_type>);
-
-    public:
-        using value_type = typename SingleLevelSet::value_type;
-
-        template <typename Arg, HTKind ht_kind = HTKind::Unknown>
-        auto ALWAYS_INLINE insert(Arg && arg)
-        {
-            if constexpr (ht_kind == HTKind::TwoLevel)
-                asTwoLevel().insert(std::forward<Arg>(arg));
-            else
-                asSingleLevel().insert(std::forward<Arg>(arg));
-        }
-
-        auto merge(const Set & other, ThreadPool * thread_pool = nullptr)
-        {
-            if (isSingleLevel() && other.isTwoLevel())
-                convertToTwoLevel();
-
-            if (isSingleLevel())
-            {
-                asSingleLevel().merge(other.asSingleLevel());
-            }
-            else
-            {
-                auto & lhs = asTwoLevel();
-                const auto rhs_ptr = other.getTwoLevelSet();
-                const auto & rhs = *rhs_ptr;
-                if (!thread_pool)
-                {
-                    for (size_t i = 0; i < rhs.NUM_BUCKETS; ++i)
-                        lhs.impls[i].merge(rhs.impls[i]);
-                }
-                else
-                {
-                    auto next_bucket_to_merge = std::make_shared<std::atomic_uint32_t>(0);
-
-                    auto thread_func = [&lhs, &rhs, next_bucket_to_merge, thread_group = CurrentThread::getGroup()]()
-                    {
-                        if (thread_group)
-                            CurrentThread::attachToIfDetached(thread_group);
-                        setThreadName("UniqExactMerger");
-
-                        while (true)
-                        {
-                            const auto bucket = next_bucket_to_merge->fetch_add(1);
-                            if (bucket >= rhs.NUM_BUCKETS)
-                                return;
-                            lhs.impls[bucket].merge(rhs.impls[bucket]);
-                        }
-                    };
-
-                    for (size_t i = 0; i < thread_pool->getMaxThreads(); ++i)
-                        thread_pool->scheduleOrThrowOnError(thread_func);
-                    thread_pool->wait();
-                }
-            }
-        }
-
-        void read(ReadBuffer & in)
-        {
-            asSingleLevel().read(in);
-        }
-
-        void write(WriteBuffer & out) const
-        {
-            if (isSingleLevel())
-                asSingleLevel().write(out);
-            else
-                getTwoLevelSet()->writeAsSingleLevel(out);
-        }
-
-        size_t size() const { return isSingleLevel() ? asSingleLevel().size() : asTwoLevel().size(); }
-
-        std::shared_ptr<TwoLevelSet> getTwoLevelSet() const
-        {
-            return two_level_set ? two_level_set : std::make_shared<TwoLevelSet>(asSingleLevel());
-        }
-
-        void convertToTwoLevel()
-        {
-            two_level_set = getTwoLevelSet();
-            single_level_set.clear();
-        }
-
-        bool isSingleLevel() const { return !two_level_set; }
-        bool isTwoLevel() const { return !!two_level_set; }
-
-    private:
-        SingleLevelSet & asSingleLevel() { return single_level_set; }
-        const SingleLevelSet & asSingleLevel() const { return single_level_set; }
-
-        TwoLevelSet & asTwoLevel() { return *two_level_set; }
-        const TwoLevelSet & asTwoLevel() const { return *two_level_set; }
-
-        SingleLevelSet single_level_set;
-        std::shared_ptr<TwoLevelSet> two_level_set;
-    };
 }
 
 
@@ -221,7 +108,7 @@ struct AggregateFunctionUniqExactData
     /// When creating, the hash table must be small.
     using SingleLevelSet = HashSet<Key, HashCRC32<Key>, HashTableGrower<4>, HashTableAllocatorWithStackMemory<sizeof(Key) * (1 << 4)>>;
     using TwoLevelSet = TwoLevelHashSet<Key, HashCRC32<Key>>;
-    using Set = detail::Set<SingleLevelSet, TwoLevelSet>;
+    using Set = UniqExactSet<SingleLevelSet, TwoLevelSet>;
 
     Set set;
 
@@ -237,7 +124,7 @@ struct AggregateFunctionUniqExactData<String>
     /// When creating, the hash table must be small.
     using SingleLevelSet = HashSet<Key, UInt128TrivialHash, HashTableGrower<3>, HashTableAllocatorWithStackMemory<sizeof(Key) * (1 << 3)>>;
     using TwoLevelSet = TwoLevelHashSet<Key, UInt128TrivialHash>;
-    using Set = detail::Set<SingleLevelSet, TwoLevelSet>;
+    using Set = UniqExactSet<SingleLevelSet, TwoLevelSet>;
 
     Set set;
 
@@ -298,13 +185,14 @@ template <typename T> struct AggregateFunctionUniqTraits
 template <typename T, typename Data, bool is_variadic = false, bool is_exact = false, bool argument_is_tuple = false>
 struct Adder
 {
-    template <HTKind ht_kind = HTKind::Unknown>
+    template <bool use_single_level_hash_table = true>
     static void ALWAYS_INLINE add(Data & data, const IColumn ** columns, size_t num_args, size_t row_num)
     {
         if constexpr (is_variadic)
         {
             if constexpr (IsUniqExactData<Data>::value)
-                data.set.template insert<T, ht_kind>(UniqVariadicHash<is_exact, argument_is_tuple>::apply(num_args, columns, row_num));
+                data.set.template insert<T, use_single_level_hash_table>(
+                    UniqVariadicHash<is_exact, argument_is_tuple>::apply(num_args, columns, row_num));
             else
                 data.set.insert(T{UniqVariadicHash<is_exact, argument_is_tuple>::apply(num_args, columns, row_num)});
             return;
@@ -331,8 +219,8 @@ struct Adder
             const auto & column = *columns[0];
             if constexpr (!std::is_same_v<T, String>)
             {
-                using Type = decltype(std::declval<const ColumnVector<T> &>().getData()[row_num]);
-                data.set.template insert<Type, ht_kind>(assert_cast<const ColumnVector<T> &>(column).getData()[row_num]);
+                data.set.template insert<const T &, use_single_level_hash_table>(
+                    assert_cast<const ColumnVector<T> &>(column).getData()[row_num]);
             }
             else
             {
@@ -343,7 +231,7 @@ struct Adder
                 hash.update(value.data, value.size);
                 hash.get128(key);
 
-                data.set.template insert<const UInt128 &, ht_kind>(key);
+                data.set.template insert<const UInt128 &, use_single_level_hash_table>(key);
             }
         }
 #if USE_DATASKETCHES
@@ -357,16 +245,15 @@ struct Adder
 
     static void add(Data & data, const IColumn ** columns, size_t num_args, size_t row_begin, size_t row_end, const char8_t * flags, const UInt8 * null_map)
     {
-        HTKind ht_kind = HTKind::Unknown;
+        bool use_single_level_hash_table = true;
         if constexpr (IsUniqExactData<Data>::value)
-            ht_kind = data.set.isSingleLevel() ? HTKind::SingleLevel : HTKind::TwoLevel;
+            if (data.set.isTwoLevel())
+                use_single_level_hash_table = false;
 
-        if (ht_kind == HTKind::Unknown)
-            add<HTKind::Unknown>(data, columns, num_args, row_begin, row_end, flags, null_map);
-        else if (ht_kind == HTKind::SingleLevel)
-            add<HTKind::SingleLevel>(data, columns, num_args, row_begin, row_end, flags, null_map);
+        if (use_single_level_hash_table)
+            add<true>(data, columns, num_args, row_begin, row_end, flags, null_map);
         else
-            add<HTKind::TwoLevel>(data, columns, num_args, row_begin, row_end, flags, null_map);
+            add<false>(data, columns, num_args, row_begin, row_end, flags, null_map);
 
         if constexpr (IsUniqExactData<Data>::value)
         {
@@ -376,21 +263,22 @@ struct Adder
     }
 
 private:
-    template <HTKind ht_kind>
-    static void add(Data & data, const IColumn ** columns, size_t num_args, size_t row_begin, size_t row_end, const char8_t * flags, const UInt8 * null_map)
+    template <bool use_single_level_hash_table>
+    static void
+    add(Data & data, const IColumn ** columns, size_t num_args, size_t row_begin, size_t row_end, const char8_t * flags, const UInt8 * null_map)
     {
         if (!flags)
         {
             if (!null_map)
             {
                 for (size_t row = row_begin; row < row_end; ++row)
-                    add<ht_kind>(data, columns, num_args, row);
+                    add<use_single_level_hash_table>(data, columns, num_args, row);
             }
             else
             {
                 for (size_t row = row_begin; row < row_end; ++row)
                     if (!null_map[row])
-                        add<ht_kind>(data, columns, num_args, row);
+                        add<use_single_level_hash_table>(data, columns, num_args, row);
             }
         }
         else
@@ -399,13 +287,13 @@ private:
             {
                 for (size_t row = row_begin; row < row_end; ++row)
                     if (flags[row])
-                        add<ht_kind>(data, columns, num_args, row);
+                        add<use_single_level_hash_table>(data, columns, num_args, row);
             }
             else
             {
                 for (size_t row = row_begin; row < row_end; ++row)
                     if (!null_map[row] && flags[row])
-                        add<ht_kind>(data, columns, num_args, row);
+                        add<use_single_level_hash_table>(data, columns, num_args, row);
             }
         }
     }
